refactor: update primitive types
