+++ primitives/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "primitives"
+version = "0.0.0"
+authors = ['AtScale <https://github.com/atscaletech>']
+edition = "2021"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+serde = { version = "1.0.124", optional = true, default-features = false }
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = [
+  "derive",
+] }
+scale-info = { version = "1.0", default-features = false, features = [
+  "derive",
+] }
+sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16" }
+sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16" }
+sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16" }
+sp-io = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16" }
+
+[features]
+default = ["std"]
+std = [
+  "codec/std",
+  "scale-info/std",
+  "serde",
+  "sp-std/std",
+  "sp-core/std",
+  "sp-runtime/std",
+  "sp-io/std"
+]
